refactor(grpc): unify SearchResult precision to f32 across proto/core
Eliminates the f32↔f64 round-trip that every layer boundary used to
perform because proto/vectorizer.proto declared `double score` while
`crate::models::SearchResult::score` is `f32`. The narrowing at the
proto→models hop silently reordered near-tie results. cluster.proto
was already `float`; vectorizer.proto now matches.

Wire-breaking (bumped in CHANGELOG [Unreleased] Breaking):
- SearchResult.score: double → float
- HybridSearchResult.{hybrid_score,dense_score,sparse_score}: double → float

New conversions (both ways, explicit, greppable):
- impl From<&models::SearchResult> for proto::SearchResult (pre-existing, updated to drop the `as f64`)
- impl From<proto::SearchResult> for models::SearchResult (new)
- impl From<proto::HybridSearchResult> for models::SearchResult (new;
  flattens hybrid_score → score, preserves dense/sparse components)

Ranker rename: `src/search/advanced_search::SearchResult` →
`ScoredDocument`. It's an augmented shape (document_id/title/snippet/
score_breakdown/highlights) unrelated to the HNSW scoring shape; the
shared name was an accident-prone footgun. 13 call sites renamed; no
external consumers (`pub use advanced_search::*;` in search/mod.rs has
zero crate-external users of the name).

Build fix: added explicit `cargo:rerun-if-changed` for
proto/vectorizer.proto and proto/cluster.proto in build.rs. Without
them a proto edit silently failed to regenerate src/grpc/vectorizer.rs
on the next `cargo build` (observed during this task — needed to
`touch build.rs` to force the regeneration).

Tests: 4 new unit tests in src/grpc/conversions.rs — bit-exact f32
round-trip, ordering preservation through 3 near-tied scores,
HybridSearchResult → SearchResult field mapping, empty-field None
handling.

Docs: new docs/api/data-model.md enumerates the canonical type, its
layer-boundary conversions, and a "not SearchResult" table documenting
ScoredDocument, SDK SearchResult, and benchmark SearchResults.

Quality gates: clippy 0 warnings, fmt clean, 1120/1120 lib (+4),
776/776 integration, both CI gate scripts clean.

+++ src/grpc/server.rs
@@ -501,9 +501,10 @@ impl VectorizerService for VectorizerGrpcService {
             .iter()
             .map(|r| vectorizer::HybridSearchResult {
                 id: r.id.clone(),
-                hybrid_score: r.score as f64,
-                dense_score: r.dense_score.unwrap_or(0.0) as f64,
-                sparse_score: r.sparse_score.unwrap_or(0.0) as f64,
+                // proto fields are now `float` (f32). See phase2_unify-search-result-type.
+                hybrid_score: r.score,
+                dense_score: r.dense_score.unwrap_or(0.0),
+                sparse_score: r.sparse_score.unwrap_or(0.0),
                 vector: r.vector.as_ref().map(|v| v.clone()).unwrap_or_default(),
                 payload: r
                     .payload
